Move vanilla editor stuff into its own file.
